backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.1
SQLAlchemy==2.0.38
PyMySQL==1.1.1
aiomysql==0.2.0
